+++ R/p_value_betwithin.R
@@ -7,25 +7,27 @@
 #' @param dof Degrees of Freedom.
 #' @inheritParams ci.default
 #'
-#' @details \subsection{Small Sample Cluster corrected Degrees of Freedom}{
+#' @details
+#' ## Small Sample Cluster corrected Degrees of Freedom
 #' Inferential statistics (like p-values, confidence intervals and
 #' standard errors) may be biased in mixed models when the number of clusters
 #' is small (even if the sample size of level-1 units is high). In such cases
 #' it is recommended to approximate a more accurate number of degrees of freedom
-#' for such inferential statistics (see \cite{Li and Redden 2015}). The
+#' for such inferential statistics (see _Li and Redden 2015_). The
 #' *Between-within* denominator degrees of freedom approximation is
 #' recommended in particular for (generalized) linear mixed models with repeated
 #' measurements (longitudinal design). `dof_betwithin()` implements a heuristic
 #' based on the between-within approach. **Note** that this implementation
-#' does not return exactly the same results as shown in \cite{Li and Redden 2015},
+#' does not return exactly the same results as shown in _Li and Redden 2015_,
 #' but similar.
-#' }
-#' \subsection{Degrees of Freedom for Longitudinal Designs (Repeated Measures)}{
+#'
+#' ## Degrees of Freedom for Longitudinal Designs (Repeated Measures)
 #' In particular for repeated measure designs (longitudinal data analysis),
 #' the *between-within* heuristic is likely to be more accurate than simply
 #' using the residual or infinite degrees of freedom, because `dof_betwithin()`
-#' returns different degrees of freedom for within-cluster and between-cluster effects.
-#' }
+#' returns different degrees of freedom for within-cluster and between-cluster
+#' effects.
+#'
 #' @seealso `dof_betwithin()` is a small helper-function to calculate approximated
 #' degrees of freedom of model parameters, based on the "between-within" heuristic.
 #'