+++ R/p_value_ml1.R
@@ -8,31 +8,32 @@
 #' @param dof Degrees of Freedom.
 #' @inheritParams ci.default
 #'
-#' @details \subsection{Small Sample Cluster corrected Degrees of Freedom}{
+#' @details
+#' ## Small Sample Cluster corrected Degrees of Freedom
 #' Inferential statistics (like p-values, confidence intervals and
 #' standard errors) may be biased in mixed models when the number of clusters
 #' is small (even if the sample size of level-1 units is high). In such cases
 #' it is recommended to approximate a more accurate number of degrees of freedom
-#' for such inferential statistics (see \cite{Li and Redden 2015}). The
+#' for such inferential statistics (see _Li and Redden 2015_). The
 #' *m-l-1* heuristic is such an approach that uses a t-distribution with
 #' fewer degrees of freedom (`dof_ml1()`) to calculate p-values
 #' (`p_value_ml1()`) and confidence intervals (`ci(method = "ml1")`).
-#' }
-#' \subsection{Degrees of Freedom for Longitudinal Designs (Repeated Measures)}{
+#'
+#' ## Degrees of Freedom for Longitudinal Designs (Repeated Measures)
 #' In particular for repeated measure designs (longitudinal data analysis),
 #' the *m-l-1* heuristic is likely to be more accurate than simply using the
 #' residual or infinite degrees of freedom, because `dof_ml1()` returns
 #' different degrees of freedom for within-cluster and between-cluster effects.
-#' }
-#' \subsection{Limitations of the "m-l-1" Heuristic}{
+#'
+#' ## Limitations of the "m-l-1" Heuristic
 #' Note that the "m-l-1" heuristic is not applicable (or at least less accurate)
 #' for complex multilevel designs, e.g. with cross-classified clusters. In such cases,
 #' more accurate approaches like the Kenward-Roger approximation (`dof_kenward()`)
 #' is recommended. However, the "m-l-1" heuristic also applies to generalized
 #' mixed models, while approaches like Kenward-Roger or Satterthwaite are limited
 #' to linear mixed models only.
-#' }
-#' @seealso `dof_ml1()` is a small helper-function to calculate approximated
+#'
+#' @seealso [`dof_ml1()`] is a small helper-function to calculate approximated
 #' degrees of freedom of model parameters, based on the "m-l-1" heuristic.
 #'
 #' @examples