+++ R/utils_model_parameters.R
@@ -340,6 +340,13 @@
     clean_params$Cleaned_Parameter[match(params$Parameter, clean_params$Parameter)],
     params$Parameter
   )
+
+  # add Group variable
+  if (!is.null(clean_params$Group) && any(nchar(clean_params$Group) > 0)) {
+    params$Group <- tryCatch(gsub("(.*): (.*)", "\\2", clean_params$Group),
+                             error = function(e) NULL)
+  }
+
   attr(params, "cleaned_parameters") <- named_clean_params
   attr(params, "pretty_names") <- named_clean_params