+++ tests/testthat/test-marginaleffects.R
@@ -48,14 +48,13 @@ test_that("deltamethod()", {
 })
 
 
-
 test_that("multiple contrasts: Issue #779", {
   mod <- lm(mpg ~ as.factor(gear) * as.factor(cyl), data = mtcars)
   cmp <- suppressWarnings(comparisons(
     mod,
     variables = c("gear", "cyl"),
     newdata = insight::get_datagrid(mod, at = c("gear", "cyl")),
-    interaction = TRUE
+    cross = TRUE
   ))
   cmp <- parameters(cmp)
   expect_true("Comparison: gear" %in% colnames(cmp))