+++ NEWS.md
@@ -1,4 +1,16 @@
-# parameters 0.19.1
+# parameters 0.20.1
+
+## Changes
+
+* `model_parameters()` now includes a `Group` column `stanreg` or `brmsfit`
+  models with random effects.
+
+## Bug fix
+
+* Fixed issue with the `print()` method for `compare_parameters()`, which
+  duplicated random effects parameters rows in some edge cases.
+
+# parameters 0.20.0
 
 ## Breaking
 
@@ -9,6 +21,17 @@
   have been removed. These were superseded by the `vcov` argument in `ci()`,
   `p_value()`, and `standard_error()`, respectively.
 
+* The `style` argument in `compare_parameters()` was renamed into `select`.
+
+## New functions
+
+* `p_function()`, to print and plot p-values and compatibility (confidence)
+  intervals for statistical models, at different levels. This allows to see
+  which estimates are most compatible with the model at various compatibility
+  levels.
+
+* `p_calibrate()`, to compute calibrated p-values.
+
 ## Changes
 
 * `model_parameters()` and `compare_parameters()` now use the unicode character
@@ -16,9 +39,14 @@
   `options(parameters_interaction = <value>)` or the argument `interaction_mark`
   to use a different character as interaction mark.
 
-* The `style` argument in `compare_parameters()`, which is used to control the
+* The `select` argument in `compare_parameters()`, which is used to control the
   table column elements, now supports an experimental glue-like syntax.
-  See this vignette _Printing Model Parameters_.
+  See this vignette _Printing Model Parameters_. Furthermore, the `select`
+  argument can also be used in the `print()` method for `model_parameters()`.
+
+* `print_html()` gets a `font_size` and `line_padding` argument to tweak the
+  appearance of HTML tables. Furthermore, arguments `select` and `column_labels`
+  are new, to customize the column layout of tables. See examples in `?display`.
 
 * Consolidation of vignettes on standardization of model parameters.